// src/features/filter/types/filter-type.ts
export enum OrderFilterKey {
  status = 'status',
  docNumber = 'docNumber',
  amount = 'amount',
  docType = 'docType',
  recipientRegistryCode = 'recipientRegistryCode',
  recipientPersonalAccount = 'recipientPersonalAccount',
  recipientCardNumber = 'recipientCardNumber',
  recipientName = 'recipientName',
  recipientBankAccount = 'recipientBankAccount',
  recipientBankName = 'recipientBankName',
  payerName = 'payerName',
  payerBankAccount = 'payerBankAccount',
  payerBankName = 'payerBankName',
  docDateFrom = 'docDateFrom',
  docDateTo = 'docDateTo',
  createdTsFrom = 'createdTsFrom',
  createdTsTo = 'createdTsTo',
  createdTs = 'createdTs',
}

export type FilterType = {
  [OrderFilterKey.docNumber]: null | string,
  [OrderFilterKey.docDateFrom]: null | string,
  [OrderFilterKey.docDateTo]: null | string,
  [OrderFilterKey.createdTsFrom]: null | string,
  [OrderFilterKey.createdTsTo]: null | string,
  [OrderFilterKey.amount]: null | number,
  [OrderFilterKey.docType]: null | string,
  [OrderFilterKey.recipientRegistryCode]: null | string,
  [OrderFilterKey.recipientPersonalAccount]: null | string,
  [OrderFilterKey.recipientCardNumber]: null | string,
  [OrderFilterKey.recipientName]: null | string,
  [OrderFilterKey.recipientBankAccount]: null | string,
  [OrderFilterKey.recipientBankName]: null | string,
  [OrderFilterKey.payerName]: null | string,
  [OrderFilterKey.payerBankAccount]: null | string,
  [OrderFilterKey.payerBankName]: null | string,
  [OrderFilterKey.status]: {
    label: string,
    value: string,
  }[] | null,
}